+++ src/components/list-page/list-card-container/ListCardContainer.scss
