+++ src/components/list-page/list-card/ListCard.scss
