+++ src/components/list-page/writers-profiles/WritersProfiles.scss
